Refactor: rearrange packages
