+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/offer/components/MuSigPriceInputBox.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.mu_sig.components.offer;
+package bisq.desktop.main.content.mu_sig.offer.components;
 
 import bisq.desktop.components.controls.MaterialTextField;
 import javafx.beans.property.StringProperty;
