+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/offer/create_offer/payment/MuSigCreateOfferPaymentController.java
@@ -33,7 +33,7 @@ import bisq.desktop.common.utils.KeyHandlerUtil;
 import bisq.desktop.common.view.Controller;
 import bisq.desktop.common.view.Navigation;
 import bisq.desktop.components.overlay.Popup;
-import bisq.desktop.main.content.mu_sig.components.offer.MuSigPaymentMethodChipButton;
+import bisq.desktop.main.content.mu_sig.offer.components.MuSigPaymentMethodChipButton;
 import bisq.desktop.navigation.NavigationTarget;
 import bisq.desktop.overlay.OverlayController;
 import bisq.i18n.Res;
