+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/offer/create_offer/payment/MuSigCreateOfferPaymentView.java
@@ -27,7 +27,7 @@ import bisq.desktop.components.containers.Spacer;
 import bisq.desktop.components.containers.WizardOverlay;
 import bisq.desktop.components.controls.AutoCompleteComboBox;
 import bisq.desktop.components.controls.BisqTooltip;
-import bisq.desktop.main.content.mu_sig.components.offer.MuSigPaymentMethodChipButton;
+import bisq.desktop.main.content.mu_sig.offer.components.MuSigPaymentMethodChipButton;
 import bisq.i18n.Res;
 import javafx.collections.ListChangeListener;
 import javafx.geometry.Insets;
