+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/offer/take_offer/amount/MuSigTakeOfferAmountController.java
@@ -25,7 +25,7 @@ import bisq.desktop.ServiceProvider;
 import bisq.desktop.common.Browser;
 import bisq.desktop.common.utils.KeyHandlerUtil;
 import bisq.desktop.common.view.Controller;
-import bisq.desktop.main.content.mu_sig.components.offer.amount_selection.MuSigAmountSelectionController;
+import bisq.desktop.main.content.mu_sig.offer.components.amount_selection.MuSigAmountSelectionController;
 import bisq.i18n.Res;
 import bisq.offer.Direction;
 import bisq.offer.amount.OfferAmountUtil;
