+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/offer/take_offer/payment/MuSigTakeOfferPaymentView.java
@@ -26,7 +26,7 @@ import bisq.desktop.components.containers.Spacer;
 import bisq.desktop.components.containers.WizardOverlay;
 import bisq.desktop.components.controls.AutoCompleteComboBox;
 import bisq.desktop.components.controls.BisqTooltip;
-import bisq.desktop.main.content.mu_sig.components.offer.MuSigPaymentMethodChipButton;
+import bisq.desktop.main.content.mu_sig.offer.components.MuSigPaymentMethodChipButton;
 import bisq.i18n.Res;
 import bisq.account.payment_method.PaymentMethodSpec;
 import javafx.geometry.Insets;
