+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/offer/take_offer/review/MuSigTakeOfferReviewController.java
@@ -32,8 +32,8 @@ import bisq.desktop.common.threading.UIScheduler;
 import bisq.desktop.common.threading.UIThread;
 import bisq.desktop.common.view.Controller;
 import bisq.desktop.components.overlay.Popup;
-import bisq.desktop.main.content.mu_sig.components.offer.MuSigPriceInput;
-import bisq.desktop.main.content.mu_sig.components.offer.MuSigReviewDataDisplay;
+import bisq.desktop.main.content.mu_sig.offer.components.MuSigPriceInput;
+import bisq.desktop.main.content.mu_sig.offer.components.MuSigReviewDataDisplay;
 import bisq.desktop.navigation.NavigationTarget;
 import bisq.i18n.Res;
 import bisq.mu_sig.MuSigService;
