+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/offer/take_offer/review/MuSigTakeOfferReviewView.java
@@ -25,8 +25,8 @@ import bisq.desktop.common.view.View;
 import bisq.desktop.components.containers.WizardOverlay;
 import bisq.desktop.components.controls.BisqTooltip;
 import bisq.desktop.components.controls.TextFlowUtils;
-import bisq.desktop.main.content.mu_sig.components.trade.MuSigProtocolWaitingAnimation;
-import bisq.desktop.main.content.mu_sig.components.trade.MuSigProtocolWaitingState;
+import bisq.desktop.main.content.mu_sig.trade.components.MuSigProtocolWaitingAnimation;
+import bisq.desktop.main.content.mu_sig.trade.components.MuSigProtocolWaitingState;
 import bisq.i18n.Res;
 import de.jensd.fx.fontawesome.AwesomeIcon;
 import javafx.geometry.HPos;
