+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/trade/open_trades/trade_state/states/MuSigState1aSetupDepositTx.java
@@ -20,8 +20,8 @@ package bisq.desktop.main.content.mu_sig.trade.open_trades.trade_state.states;
 import bisq.chat.mu_sig.open_trades.MuSigOpenTradeChannel;
 import bisq.desktop.ServiceProvider;
 import bisq.desktop.components.controls.WrappingText;
-import bisq.desktop.main.content.mu_sig.components.trade.MuSigWaitingAnimation;
-import bisq.desktop.main.content.mu_sig.components.trade.MuSigWaitingState;
+import bisq.desktop.main.content.mu_sig.trade.components.MuSigWaitingAnimation;
+import bisq.desktop.main.content.mu_sig.trade.components.MuSigWaitingState;
 import bisq.i18n.Res;
 import bisq.trade.mu_sig.MuSigTrade;
 import javafx.scene.layout.HBox;
