+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/trade/open_trades/trade_state/states/MuSigState1bWaitForDepositTxConfirmation.java
@@ -29,8 +29,8 @@ import bisq.desktop.common.threading.UIThread;
 import bisq.desktop.components.controls.MaterialTextField;
 import bisq.desktop.components.controls.WrappingText;
 import bisq.desktop.components.controls.validator.BitcoinTransactionValidator;
-import bisq.desktop.main.content.mu_sig.components.trade.MuSigWaitingAnimation;
-import bisq.desktop.main.content.mu_sig.components.trade.MuSigWaitingState;
+import bisq.desktop.main.content.mu_sig.trade.components.MuSigWaitingAnimation;
+import bisq.desktop.main.content.mu_sig.trade.components.MuSigWaitingState;
 import bisq.i18n.Res;
 import bisq.trade.mu_sig.MuSigTrade;
 import de.jensd.fx.fontawesome.AwesomeIcon;
