+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/trade/open_trades/trade_state/states/MuSigTradeCompletedTable.java
@@ -26,8 +26,8 @@ import bisq.desktop.components.controls.BisqMenuItem;
 import bisq.desktop.components.controls.BitcoinAmountDisplay;
 import bisq.desktop.components.controls.WrappingText;
 import bisq.desktop.main.content.components.UserProfileDisplay;
-import bisq.desktop.main.content.mu_sig.components.trade.MuSigWaitingAnimation;
-import bisq.desktop.main.content.mu_sig.components.trade.MuSigWaitingState;
+import bisq.desktop.main.content.mu_sig.trade.components.MuSigWaitingAnimation;
+import bisq.desktop.main.content.mu_sig.trade.components.MuSigWaitingState;
 import bisq.i18n.Res;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
